+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added `Wizaplace\SDK\Vendor\Order\Order::getDetails` 
 - Added `Wizaplace\SDK\Vendor\Order\OrderService::setOrderDetails`
 - Added `alt` property to `\Wizaplace\SDK\Cms\Banner`
+- Added `Wizaplace\SDK\Pim\Product\ProductService::addAttachments`
+- Added `Wizaplace\SDK\Pim\Product\ProductService::removeAttachment`
+- Added `Wizaplace\SDK\Catalog\CatalogService::getProductAttachment`
 
 ### Bugfixes